refactor : refactor enum package
