+++ src/main/java/lotto/model/domain/Enum/Place.java
@@ -1,4 +1,4 @@
-package lotto.model.domain;
+package lotto.model.domain.Enum;
 
 public enum Place {
     // 등수
