+++ src/main/java/lotto/model/domain/verifier/LottoWinVerifier.java
@@ -1,8 +1,10 @@
-package lotto.model.domain;
+package lotto.model.domain.verifier;
 
 import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
+import lotto.model.domain.Enum.Place;
+import lotto.model.domain.LottoAndBonusNumber;
 
 public class LottoWinVerifier {
