+++ src/test/java/lotto/model/domain/InputVerifierTest.java
@@ -1,6 +1,7 @@
 package lotto.model.domain;
 
 
+import lotto.model.domain.verifier.InputVerifier;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
